Quite a lot of changes in the main java file.
- camera_start() is altered and renamed to safeCameraOpen()
- Trying to make the Capture button works, but still not successful yet.
- Tried to eliminate Start button because the camera should start as
soon as the app starts anyway, but not successful yet either.

Also, I would like to try and eliminate the need of Start button if
possible, but strangely enough, if I put the safeCameraOpen() inside the
onCreate function, the surfaceView does not show anything on-screen.

But if I do the safeCameraOpen() function inside a button click callback
function, the surfaceView does work fine.